move other filemngr modules to mpp
