+++ components/filemngr/listing/impl/src/androidMain/kotlin/com/flipperdevices/filemanager/listing/impl/composable/NoFilesComposablePreview.kt
@@ -1,0 +1,13 @@
+package com.flipperdevices.filemanager.listing.impl.composable
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.flipperdevices.core.ui.theme.FlipperThemeInternal
+
+@Preview
+@Composable
+private fun NoFilesComposablePreview() {
+    FlipperThemeInternal {
+        NoFilesComposable(onUploadFilesClick = {})
+    }
+}
