+++ components/filemngr/listing/impl/src/androidMain/kotlin/com/flipperdevices/filemanager/listing/impl/composable/appbar/OrangeAppBarTestPreview.kt
@@ -1,0 +1,17 @@
+package com.flipperdevices.filemanager.listing.impl.composable.appbar
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.flipperdevices.core.ui.theme.FlipperThemeInternal
+
+@Preview
+@Composable
+private fun OrangeAppBarTestPreview() {
+    FlipperThemeInternal {
+        CloseSelectionAppBar(
+            onClose = {},
+            onDeselectAll = {},
+            onSelectAll = {}
+        )
+    }
+}
