+++ components/filemngr/listing/impl/src/androidMain/kotlin/com/flipperdevices/filemanager/listing/impl/composable/modal/BottomSheetOptionsPreview.kt
@@ -1,0 +1,24 @@
+package com.flipperdevices.filemanager.listing.impl.composable.modal
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.flipperdevices.bridge.connection.feature.storage.api.model.FileType
+import com.flipperdevices.core.ui.theme.FlipperThemeInternal
+import okio.Path.Companion.toPath
+
+@Preview
+@Composable
+private fun BottomSheetOptionsPreview() {
+    FlipperThemeInternal {
+        BottomSheetOptionsContent(
+            fileType = FileType.DIR,
+            path = "some_file.ir".toPath(),
+            onExport = {},
+            onDelete = {},
+            onCopyTo = {},
+            onRename = {},
+            onSelect = {},
+            onMoveTo = {}
+        )
+    }
+}
